+++ frontend/form-builder/src/components/FormSettings/FormSettings.module.css
@@ -3,13 +3,13 @@
 }
 
 .noSelectedItem {
-  height: 85%;
+  height: 80%;
   display: flex;
   flex-direction: column;
   justify-content: center;
 }
 
-.SettingsContent {
+.settingsContent {
   display: flex;
   flex-direction: column;
   padding: var(--fds-spacing-4);
@@ -22,11 +22,13 @@
   scrollbar-color: var(--fds-semantic-surface-neutral-inverted) transparent;
 }
 
-.buttons {
+.buttonContainer {
   display: flex;
   flex-direction: column;
   gap: var(--fds-spacing-2);
-  padding-top: var(--fds-spacing-3);
+  padding-top: var(--fds-spacing-2);
+  margin-top: auto;
+  margin-bottom: var(--fds-spacing-2);
 }
 
 .modalWindow {
