+++ frontend/form-builder/src/components/Toolbar/DraggableItem/DraggableItem.module.css
@@ -12,7 +12,7 @@
 
 .toolbarIcon {
   display: flex;
-  font-size: 3rem;
+  font-size: 2.5rem;
 }
 
 @media (max-width: 768px) {
